+++ src/libutil/parse_ini.c
@@ -170,7 +170,7 @@ parse_ini_open(const char *fname)
 		line_len = strlen(line);
 		while ((line[i] != '\0') && (isblank(line[i])))
 			i++;
-		if (i == line_len - 1) {
+		if (i == (int)line_len - 1) {
 			line[0] = '\0';
 		}
